feat: move auth form login in custom hook
