+++ src/components/AuthForm/AuthForm.tsx
@@ -1,65 +1,12 @@
-import { useState } from 'react';
-
 import { CustomInput } from '../CustomInput/CustomInput';
 import { LoadingSpinner } from '../Icons/LoadingSpinner/LoadingSpinner';
 import { ValidateMessage } from '../ValidateMessage/ValidateMessage';
 
-import { emulateServerResponse } from '@/service/AuthService';
-import { isValidEmail, isValidPassword, isValidUser } from '@/utils/utils';
-import { TypeUser } from '@/types/types';
+import { useForm } from '@/hooks/useForm';
 import classes from './AuthForm.module.scss';
 
 export const AuthForm = () => {
-  const [inputs, setInputs] = useState<TypeUser>({
-    email: '',
-    password: '',
-  });
-
-  const [isValidInput, setIsValidInput] = useState({
-    email: true,
-    password: true,
-  });
-
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleChangeInput = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = event.target;
-    setInputs((state) => ({
-      ...state,
-      [name]: value,
-    }));
-    if (name === 'email') {
-      const testEmail = isValidEmail(value);
-      setIsValidInput((state) => ({
-        ...state,
-        [name]: testEmail,
-      }));
-    }
-    if (name === 'password') {
-      const testPassword = isValidPassword(value);
-      setIsValidInput((state) => ({
-        ...state,
-        [name]: testPassword,
-      }));
-    }
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    if (isValidUser(inputs)) {
-      try {
-        const result = await emulateServerResponse(inputs);
-        console.log(result);
-      } catch (error) {
-        console.error(error);
-      } finally {
-        setIsLoading(false);
-      }
-    } else {
-      setIsLoading(false);
-    }
-  };
+  const { isLoading, handleChangeInput, handleSubmit, isValidInput } = useForm();
 
   return (
     <form className={classes.form} onSubmit={(e) => void handleSubmit(e)}>
